backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
supabase==2.4.2
groq==0.9.0
google-generativeai==0.8.3
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.1
python-multipart==0.0.9
beautifulsoup4==4.12.3
lxml==5.2.2
